:art: Fix : dto 클래스와 model 클래스에서 id 타입 Long으로 변경 -  dto 클래스와 model 클래스에서 id 타입을 Long으로 변경
Related to: #153

+++ AOS/app/src/main/java/com/softeer/team6four/data/remote/charger/model/BottomSheetChargerModel.kt
@@ -2,7 +2,7 @@ package com.softeer.team6four.data.remote.charger.model
 
 data class BottomSheetChargerModel(
     val address: String,
-    val chargerId: Int,
+    val chargerId: Long,
     val feePerHour: String,
     val nickname: String,
     val speedType: String
